Release v0.4.0

+++ src/analyze/components/stats.rs
@@ -6,7 +6,7 @@
 //!
 //! The v1 scope is deliberately small: `area`, axis-aligned bounding
 //! box (as inclusive min/max coordinates), and `sum_x` / `sum_y` for
-//! the integer centroid. Higher-order moments, perimeter, holes,
+//! the integer centroid. Higher-order moments, boundary pixels, holes,
 //! Euler number, orientation, convexity, and intensity measurements
 //! are deferred to follow-up work.
 
@@ -150,7 +150,7 @@ pub(super) mod sink {
         /// boundary flag for this pixel, meaningful only when
         /// [`NEEDS_BOUNDARY`](StatsSink::NEEDS_BOUNDARY) is `true`
         /// (otherwise always `false`).
-        fn record(&mut self, compact_label: u64, first: bool, at: Coordinate, is_boundary: bool);
+        fn record(&mut self, compact_label: u32, first: bool, at: Coordinate, is_boundary: bool);
     }
 
     /// Sink that drops every record. Compiles down to no work.
@@ -158,7 +158,14 @@ pub(super) mod sink {
 
     impl StatsSink for NoStats {
         #[inline(always)]
-        fn record(&mut self, _compact_label: u64, _first: bool, _at: Coordinate, _is_boundary: bool) {}
+        fn record(
+            &mut self,
+            _compact_label: u32,
+            _first: bool,
+            _at: Coordinate,
+            _is_boundary: bool,
+        ) {
+        }
     }
 
     /// Sink that accumulates per-component stats into a `Vec` indexed
@@ -172,12 +179,12 @@ pub(super) mod sink {
         // do not need the boundary flag, so the engine's neighbour-check is
         // const-folded away and this path is unchanged.
         #[inline]
-        fn record(&mut self, compact_label: u64, first: bool, at: Coordinate, _is_boundary: bool) {
+        fn record(&mut self, compact_label: u32, first: bool, at: Coordinate, _is_boundary: bool) {
             if first {
                 // Compact labels are dense `1..=label_count`, so each
                 // new label appears exactly once with `first = true`
                 // and `compact_label == out.len() + 1`.
-                debug_assert_eq!(self.out.len() as u64, compact_label - 1);
+                debug_assert_eq!(self.out.len(), (compact_label - 1) as usize);
                 self.out.push(ComponentStats::from_seed(at));
             } else {
                 self.out[(compact_label - 1) as usize].extend(at);
@@ -186,9 +193,9 @@ pub(super) mod sink {
     }
 
     /// Sink that accumulates per-component [`BlobMeasurements`] (moments
-    /// and perimeter) into a `Vec` indexed by `compact_label - 1`. Sets
+    /// and boundary count) into a `Vec` indexed by `compact_label - 1`. Sets
     /// `NEEDS_BOUNDARY = true` so the engine runs the per-pixel boundary
-    /// check that feeds the perimeter count.
+    /// check that feeds the boundary count.
     pub(crate) struct WithMeasurements<'a> {
         pub(crate) out: &'a mut Vec<BlobMeasurements>,
     }
@@ -197,9 +204,9 @@ pub(super) mod sink {
         const NEEDS_BOUNDARY: bool = true;
 
         #[inline]
-        fn record(&mut self, compact_label: u64, first: bool, at: Coordinate, is_boundary: bool) {
+        fn record(&mut self, compact_label: u32, first: bool, at: Coordinate, is_boundary: bool) {
             if first {
-                debug_assert_eq!(self.out.len() as u64, compact_label - 1);
+                debug_assert_eq!(self.out.len(), (compact_label - 1) as usize);
                 self.out.push(BlobMeasurements::from_seed(at, is_boundary));
             } else {
                 self.out[(compact_label - 1) as usize].extend(at, is_boundary);
